Use index as main file inside component folders
