--- app/src/Homepage/ResumeModal/index.js
@@ -1,1 +1,0 @@
-export { default } from './ResumeModal';
