--- app/src/Homepage/index.js
@@ -1,1 +1,0 @@
-export { default } from './Homepage';
