--- app/src/TradePage/OpenPositions/index.js
@@ -1,1 +1,0 @@
-export { default } from './OpenPositions';
