--- app/src/TradePage/StockTwitsWidget/index.js
@@ -1,1 +1,0 @@
-export { default } from './StockTwitsWidget';
