+++ app/src/TradePage/StockTwitsWidget/index.tsx
