--- app/src/TradePage/TradeHistory/index.js
@@ -1,1 +1,0 @@
-export { default } from './TradeHistory';
