--- app/src/shared/components/FadeIn/index.js
@@ -1,1 +1,0 @@
-export { default } from './FadeIn';
