--- app/src/shared/components/Footer/index.js
@@ -1,1 +1,0 @@
-export { default } from './Footer';
