--- app/src/shared/components/Header/Text/index.js
@@ -1,1 +1,0 @@
-export { default } from './Text';
