+++ app/src/shared/components/Header/Text/index.tsx
