--- app/src/shared/components/SocialButtonGroup/SocialButton/index.js
@@ -1,1 +1,0 @@
-export { default } from './SocialButton';
