--- app/src/shared/components/SocialButtonGroup/index.js
@@ -1,1 +1,0 @@
-export { default } from './SocialButtonGroup';
