--- app/src/shared/constants/Colors/Colors.js
@@ -1,6 +1,0 @@
-const COLORS = {
-  white: '#ededed',
-  grey: '#c1c1c1',
-};
-
-export default COLORS;
