+++ app/src/shared/constants/Colors/index.js
@@ -1,1 +1,6 @@
-export { default } from './Colors';
+const COLORS = {
+  white: '#ededed',
+  grey: '#c1c1c1',
+};
+
+export default COLORS;
