--- app/src/shared/constants/Common/Common.js
@@ -1,30 +1,0 @@
-const reactIsInDevelomentMode = () => process.env.NODE_ENV === 'development';
-
-const OWNER = {
-  firstName: 'Mo',
-  lastName: 'Beigi',
-  fullName: () => [OWNER.firstName, OWNER.lastName].join(' '),
-};
-
-const ANALYTICS = {
-  gaTrackingId: 'UA-35338966-1',
-};
-
-const WEBSITE = {
-  baseURL: reactIsInDevelomentMode() ? 'http://localhost:3001' : 'https://mobeigi.com',
-  titlePrefix: 'Mo Beigi: ',
-  foundingYear: 2012,
-};
-
-const SERVER = {
-  baseURL: reactIsInDevelomentMode() ? 'http://localhost:3000' : 'https://mobeigi.com/api',
-};
-
-const COMMON = {
-  ANALYTICS,
-  OWNER,
-  SERVER,
-  WEBSITE,
-};
-
-export default COMMON;
