+++ app/src/shared/utils/TargetAwareLink/index.tsx
